pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wizctl"
version = "0.1.0"
description = "Control Philips WiZ smart lights from CLI or REST API"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Bharat Sharma" }
]
keywords = ["wiz", "smart-lights", "home-automation", "philips", "cli", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]

dependencies = [
    "pywizlight>=0.5.14",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
wizctl = "wizctl.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/wizctl"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py39"]

